Force numerical context for expires_in.
Depending on where the expires_in value comes from, and depending on the
JSON encoder, it is possible that the returned value is put inside
quotes as in :

  ...,
  "expires_in":"3600",
  ...,

Examples in the RFC6749 (e.g. sections 4.1.4, 4.3.3, 4.4.3 and 5.1) show
that the value should not be enclosed in quotes.

I got caught by this while working on a provider extracting expiration out of
a database against a really picky client and this patched solved my
problem.