Fix some random test errors by sorting hash keys.
I experienced some random errors with out-of-order fields and values.
The content was effectively the same, but in a different order.

When outputing the content of a hash into a string-like format (XML
and/or URL) sort the keys of the hash before processing so that we can
get a reproducible result that can easily be tested.